+++ ProjP3.Domain/InterfaceRepositories/IRepository.cs
@@ -9,10 +9,10 @@ namespace ProjP3.Domain.InterfaceRepositories
     public interface IRepository<T> where T : class
     {
         Task<bool> SaveAllAsync();
-        Task<bool> DeleteAsync(ulong id);
-        Task<bool> ExistsAsync(ulong id);
+        Task<bool> DeleteAsync(int id);
+        Task<bool> ExistsAsync(int id);
         Task<List<T>> GetAllAsync();
-        Task<T> GetByIdAsync(ulong id);
+        Task<T> GetByIdAsync(int id);
         Task<T> AddAsync(T entity);
         Task<T> UpdateAsync(T entity);
     }
